Title: 물체 놓기, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 241121/물체 놓기/place-object.java
@@ -15,7 +15,7 @@ public class Main {
 	
 	private static int minPutV = -1;
 	private static int minPutCost = Integer.MAX_VALUE;
-	private static int sum = 0;
+	private static long sum = 0;
 	
 	private static List<Edge>[] graph;
